Doc: Show that style import needs to come first
Currently, the Styling Qt Quick Controls docs contain a code snippet
to show how to import a style.

This fix adds a comment to the code snippet to let users know that
the style import should precede all other imports. If the user imports
something else before the style import, the style does not display
correctly, namely, some of the controls follow the style, whilst some
do not.

Fixes: QTBUG-130792
Pick-to: 6.8
Change-Id: I556fc987fe7c15193aa2657fc76b61079acd72f4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>