Add drag and drop to the QtQuick Dialogs SideBar
The current SideBar implementation for the QtQuick Dialogs non-native
FileDialog has a SideBar containing supported standard paths. This patch
adds support for adding favorites to the SideBar, that can be added by
dragging a directory into the SideBar.

Each FileDialogDelegate contains a QQuickTapHandler which listens for a
long press. When long pressed, the delegate can be dragged. A long press
is used to initiate the drag as the delegates are part of a Flickable.
The drag and drop is a system drag and drop. If the dragged delegate is
a directory, the drag is started.

When the dragging enters the SideBar, a delegate appears with 'Add
Favorite' text, to suggest to the user that this is the drop area.

The context menu used for removing a favorite is a QQuickContextMenu,
which handles the context menu event internally.

Pick-to: 6.9
Change-Id: Icf60f3785522df607ed1f7420486da4318118f0f
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>