doc: Add note about assigning items directly to fillItem
Using items with no visual parent as sources causes error messages
and empty contents, both for shader effects and as fill items.

This is a particularly easy mistake to make with shapes, since the
fact that ShapePath is not a QQuickItem is not obvious. At the
least we should mention this case in the documentation and provide
a solution for it.

Pick-to: 6.8
Task-number: QTBUG-130893
Change-Id: Ibeed382a8fa35cf69e8d5a4b61b04dc5e77df217
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>