Flickable childMouseEventFilter: return early for multi-touch events
As it stood, we would only set stealMouse to false when detecting
a multi-touch event. But we would still go on checking if a
drag occurred, and perhaps even end up grabbing the mouse again.
This is against the comment that says: 'If multiple touchpoints are
within bounds, don't grab: it's probably meant for multi-touch
interaction in some child'.

This patch will therefore return early when we detect a multi-touch
event, since a multi-touch event cannot be a part of a drag gesture
that should start (or maintain) a flicking session.

Amends 5171bcd9ce99e794c1091069810314fb5457703b

Pick-to: 6.8
Task-number: QTBUG-34881
Change-Id: Ia4714b555dc0d3922cd4f5de73eaeed375d23d82
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>