qmlformat: separate reindentAndSplit's concerns
This also changes the behavior of linebreak by combining the split
sequence in a single set. Prior to this commit, linebreak position was
attempted to be detected for a set of tokens. Then, if there is a left
paranthesis in the same line, that would take the split location.

Now change this logic so that break the line after the last token that
lays on column smaller than the allowed maximum width.

Task-number: QTBUG-113590
Change-Id: Ib0e7291a139265ab9bc09be0c574704da25fd722
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>