qqmlloader: Run QQmlTypeLoader on main thread for WebAssembly platform
Multithreading on WebAssembly raises some challenges. Blocking the main
thread in the wasm instance prevents the web browser from doing work and
from processing events, which can cause deadlocks. In particular it can
block the browser from creating the web worker for the
QQmlTypeLoaderThread (which the main thread is waiting on).
This change will always schedule QQmlTypeLoader on main thread to avoid
this issue.

Fixes: QTBUG-129071
Change-Id: I57278589981b22e3e2807dd45767f75b4848c0a9
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Even Oscar Andersen <even.oscar.andersen@qt.io>