QmlDom: simplify AttachedInfo
The following 2 commits aim to simplify and prepare AttachedInfo
to be transformed to the FileLocations::Node, being responsible
only for the state it maintains.

1. merge AttachedInfoT into AttachedInfo
Remove complexity of templated structure. Currently we don't need it.
AttachedInfo is used only with FileLocations::Info

2. refactor AttachedInfo
Remove redundant functionality, move "Tree"-specific functionality
to the FileLocations namespace.
Leave in the AttachedInfo only "Node"-specific functionality.

Change-Id: I3793daa8f081f471e068c4286a94094656a0098d
Task-number: QTBUG-130710
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>