QQuickVisualTestUtils::compareImages: fail if image depth is not 32
The function would ignore the depth and read memory that it shouldn't,
which can cause hard-to-diagnose failures, and heap-buffer-overflows
when ASAN is enabled.

Mismatches in depth can happen when optimizing images with optipng,
for example, which may convert color images to grayscale.

compareImages was added in 7ffea4048321b86402a9ded5255ad16ac74d0518.

Pick-to: 6.5 6.8
Change-Id: Id83dcd79ae94088996ff770d90b2c1add910fb1b
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>