Add DAPrivate::eventTargets: shared pointer and context menu event logic
QQuickDeliveryAgentPrivate::eventTargets() finds items relevant to an
event at a position, where the position has to be found before calling
it, because QContextMenuEvent is not a pointer event. Relevance is
mainly determined by QQuickItem::contains() and QQuickItem::clipRect();
but an extra predicate can be passed along to encapsulate the extra
checks that QQuickDeliveryAgentPrivate::pointerTargets() needs to do.
It has a tri-state return value, implemented as std::optional<bool>,
to indicate whether the relevance should be definitely true, false, or
left as-is after the bounds check.

Task-number: QTBUG-67331
Pick-to: 6.9
Change-Id: I151faa655590c0b6a51b148889f18e3cde7c76f5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>