QmlCompiler: Make QQmlJSRegisterContentPrivate non-const
This is a major departure from the previous architecture. We can now
alter a QQmlJSRegisterContentPrivate after it has been created. This is
fundamentally necessary so that we can add further links to other types
as they are discovered. With an immutable QQmlJSRegisterContentPrivate
we have to clone every time which makes for long and rather meaningless
chains of links that are hard to navigate, eventually. Making the
register content non-const will in turn allow us to uphold the
immutability of QQmlJSScope. That is much more important because it can
fundamentally be used across different compilations.

Task-number: QTBUG-124670
Change-Id: Iacee705be3260da0a2bc91c6758f91c70a137065
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>