Change for QTRY_VERIFY flaky tst_qquicktextedit::overwriteMode()
Auto test tst_qquicktextedit::overwriteMode() is flaky where
textEdit->hasActiveFocus() is not always true where it should.

Modify QVERIFY for QTRY_VERIFY to give the necessary time to the XCB
thread to proceed.

Remove tst_QQuickTextEdit::overwriteMode as the test is no longer
flaky with this fix.

Remove the flakiness when testing on Opensuse 15.5

Change-Id: Iead4c1bd0781578ad80f13313acc7b8081e7337b
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>