QQuickIconImage: Replace updateIcon() with QQuickImageBase::load() override
The interface for how QQuickImageBase and QQuickImage handles loading of
the image is via the virtual QQuickImageBase::load() function. For some
reason QQuickIconImage was using its own updateIcon() callback for the
same use-case, which was causing issues.

For example, when updateIcon triggered a pixmapChange we would update the
fill mode, but doing so via QQuickImage::setFillMode() would trigger load(),
which in turn ended up in QQuickImageBase::loadEmptyUrl(), because the
way we load icons is not based on QQuickImageBasePrivate::url being set.

We now implement the icon loading in a QQuickIconImage::load() override,
which also lets us remove some of the manual updates we had in response
to e.g. the DPR changing.

A small tweak to QQuickImageBase::itemChange is needed, since it can't
assume that a DPR shouldn't trigger a load() just because the URL is
empty.

Change-Id: I6dea4bd23bb2c0b69ba5dbe2dbf7f95655aa406b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>