QtQml: Clean up qml_type_loader_thread #if'ery
Provide two separate implementations of QQmlThread, one with an actual
thread and one with a stub. Avoid handing out the mutex since that does
not exist in the stub.

Without a type loader thread we cannot force an asynchronous URL to load
synchronously. Introduce the respective exception in createQmlObject().

Parent the network reply proxy to a suitable object. If the type loader
thread is the main thread it will otherwise outlive the engine and
crash.

Since the type loader thread is the main thread in case of
!QT_CONFIG(qml_type_loader_thread) we don't have to specialize this case
when waiting for messages. We just cannot wait for messages from the
same thread we're on.

Change-Id: I91898f15a2930b8c0c261587f73d3a30eb5f8132
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>