Teach QQuickPopup about safe areas
We want users to be able to attach SafeArea to Popups directly, without
explicitly having to attach to the Popup's item. However QQuickSafeArea
can't do that directly, as QQuickPopup lives in Qt Quick Controls.

We solve this by introducing an interface for things that can have
SafeAreas attached to them.

Task-number: QTBUG-125373
Change-Id: I1610323132f318c855764b1c8759ce021218b4b4
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>