Teach QQmlPropertyToPropertyBinding about grouped properties
Tracking the source property via an explicit QObject and a core
index was missing the case where the property is a grouped/value
property.

We now pass QQmlPropertyIndex to pass core and value type indices. This
allows the source property to be a grouped property. The target property
can still not be a grouped property.

Change-Id: I90b25b7187ad6ddbf1dd5938719e74988599cddc
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>