QtQml: Fix some type loader threading issues
We need to make sure not to concurrently access the "waiting" lists in
the type loader and engine threads.

In clearCache() we just temporarily shut the type loader thread down so
that we can drop all the blobs without locking. In trimCache() we can
know that the waitingFor list is empty and therefore the blob to be
dropped cannot interfere with other blobs anymore. This requires the
status member to follow acquire/release semantics and the operation on
the waitingFor list to precede the setting of the status member. We also
avoid some nonsensical state changes in order to simplify the logic.

Finally, when detecting a cycle, we need to break the cycle before
setting the status.

To reflect the insights about trimCache and clearCache, acquire the lock
when calling trimThread from the engine, but do not acquire the lock for
clearCache.

Task-number: QTBUG-131721
Change-Id: Ib1cd974e5284f3e7e1f14dca31bddae2ea9de1b9
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>