Detect safe area binding loops
Updating a safe area's margins may result in updating bindings that
then cause the safe area margins to change again.

In some cases this is fine, because it's part of an item doing multi-
step layout, e.g. individual setPosition and setSize calls, each of
which will result in a margin update via the transform change. The
final safe area margins will then stabilize back to what they were
on the first change. However the generic binding machinery will see
this as a binding loop, and warn about it, which we don't want.

To fix this, we do custom binding loop detection for safe areas,
where we avoid the recursive emit of marginsChanged.

To catch actual binding loops, we still update margins internally,
and then check if the margins are stable when returning from the root
emit. We need a few rounds before we can conclude that the margins
have stabilized, as Qt Quick Controls adds additional levels of
bindings that need settle down.

Change-Id: Icd61edd05c2b5605ab635baac6c2488a6d444a5c
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>