Include item width/height when computing safe area bottom/right margins
We can't just naively map the bottom-right margin to the target item
when computing item-relative safe area margins, as the margins are
relative to the source item's rect, meaning we have to include the
height and width of the source item to get the coordinate to map
from.

Similarly, we have to include the target item's width/height when
then mapping from the bottom/right coordinate back to a margin.

Change-Id: If68b420bd42e038e01e6e1116863c57f761e88bc
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>