Pass transformed item on to QQuickItemChangeListener::itemTransformChanged
The item that receives the itemTransformChanged may not be the item
that causes the change. The latter is tracked by the transformedItem
argument to QQuickItemPrivate::transformChanged, so let's pass that
on as the ItemChangeData instead of a synthetic true value.

Amends 9062ba372a32536d97a6ff6a5605544a8c82f48e.

Change-Id: I6b411f62bbe1f76c06de314f8a926783de0448e3
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>