Pick up additional safe area margins when attached to window
We can't go through qmlAttachedPropertiesObject when walking
the parent item hierarchy as that will only return an existing
safe area if we ask it for the QQuickWindow's safe area, and
we've mapped that to the window's content item and attached
the safe area to that.

Instead we check for direct children of the safe area type,
which works for both the window's content item as well as
the popup's popup item.

Change-Id: Ifed146a931625f6c40475734dc5dccc1d152a885
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>