Treat safe area margins as independent when mapping from other items
We use QQuickItem mapping functions to map the top/left and bottom/right
coordinates of the safe area margins to the target item. However this
mapping doesn't treat a margin of 0 in any special way, and we still
get back a mapped QPointF relative to the target item, which depending
on the positioning and sizing of the target item may give a resulting
safe area margin, even if the original margin was 0.

To fix this we only return a mapped margin if there was an original
margin to map.

Change-Id: Id4e0d61f5a85da56a86cc01805345031073be7ec
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>