Reduce flip-flopping of safe area margins when geometry is changed
When an item changes its position or size we end up with a callback
to itemTransformChanged for every safe area attached to a descendant.
We normally want to react to the new geometry, as this might mean that
the safe areas for our item has changed.

But other intermediate items between the changed item and ourselves
may react to the geometry change by adjusting their own position or
size, which may affect the final safe area margins we compute.

These items haven't had a chance to react to the geometry change yet,
as that happens via QQuickItem::geometryChange(), after the transform
notification.

To reduce the chance of computing safe area margins that will be
invalidated again when intermediate items react to the geometry
change we hook into the geometry change event of the transformed
item, and defer our update until the geometry change event arrives.

Unfortunately this only ensures that anchors have had their effect,
as signal notifications for the item geometry properties happens
after QQuickItem::geometryChange calls notifyChangeListeners for
the geometry change.

We could theoretically extend this solution to observing the change
types of the QQuickGeometryChange, and then connecting to the
relevant change signals of the item, to update the margins even
later, but that would be a step on top of this change.

Change-Id: I1a099221d5bde9f401ea1c1888865dae8613da34
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>