TableView: ensure we position the syncView in the right direction
If a TableView has a syncView, a call to positionViewAtCell()
will instead be forwarded to the syncview, according to the
syncDirection. But as it stood, this was flipped, so that we
ended up forwarding a horizontal positioning call to a
vertically synced syncView, and vice versa.

This patch will correct this, so that the forwarded
positioning call matches the syncDirection.

Pick-to: 6.8 6.5
Fixes: QTBUG-131776
Change-Id: Ibeaf5ce3d1d18c1ef0a0184c38336860fdee67be
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>