Add Shear transform type
Shearing an item either by given angle or a relative offset
is a common operation in vector graphics.

We already have types for Rotate, Scale, etc. so this fits into
the list there. For cases where you want to animate the angles
it can be especially useful to have a transform where these
are exposed as properties.

[ChangeLog][QtQuick] Added Shear transform types for shearing
an item by a factor or angle along the x- and y-axes.

Change-Id: I6be81b40ca555a0ac0d8cb14dd81598502238a3d
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>