TableView: only create attached object when used
As it stood, we created an attached TableView delegate for
every delegates inside a TableView, even if they were never used.
This was not the intention from the start, and must have
been an oversight. Instead, we want to reduce memory overhead
for each delegate, and only create attached objects on a
per-need basis. This was also the rationale behind using
required properties over attached properties, to avoid
having to allocate so many extra attached objects (one for
each cell).

This patch will therefore avoid creating attached objects
for every delegate, and instead resolve the TableView.view
property when the attached object is actually constructed
(if ever).

Change-Id: I69528371e59d7ef4007f63029db9a46a41c40d0b
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>