qmlls: wait for exit-notification before shutdown
The LSP states that qmlls should only shutdown after having received an
"exit" notification. Previously to this patch, qmlls would shutdown
right after receiving the shutdown request, without waiting to receive
an exit notication (that is supposed to be send from the client to the
server after the shutdown request).

Amends 9aa5abbda76db44b7ead30c221675183285b420c that fixed qmlls not
being able to shutdown by making it shutdown correctly.

Add a new enum value WaitingForExit to RunStatus that describes the
state between receiving a shutdown request and before receiving the exit
notification from the client. Use it to make sure that qmlls waits for
the exit notification and that qmlls does not read again from stdin
after the exit notification was received.

Also don't exit qmlls on QLanguageServer::shutdown when the shutdown
request gets processed but rather on
QLanguageServer::exit when the exit notification has been received.

This is required for QTCREATORBUG-31897 to avoid qt creator complaining
about qmlls returning 1 when it is restarted, despite qmlls having
done a graceful shutdown.

Also make the tst_lifecycle code reusable and write tests for the
linux/macos case (where the shutdown and the exit notifications are read
in two parts) and for the windows case (where sometimes both shutdown
and exit are read in the same block from stdin).

Pick-to: 6.8 6.9
Task-number: QTCREATORBUG-31897
Change-Id: I0cd663a351027e100469e273b4bb21f47b817a49
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>