qmlformat: test linebreaks for binary operators
Also, fix the regexp handling. Prior to this commit, any character
following a divide token was assumed to be part of the regexp literal
anda eventually hits the default case in the Lexer::scanRegExp function
where it  errors out.

Restrict the assumption for space characters.scanRegExp determines
if the non-space characters are regexp literal or not.

Task-number: QTBUG-113590
Change-Id: If2b2b74661a94b3c8cf863d5e0cdc49e4b9046b4
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>