Take safe areas into account for Qt Quick Controls ToolBar and MenuBar
A ToolBar and MenuBar is commonly placed in the header/footer/menuBar
of an ApplicationWindow, where we don't do any automatic padding to
account for the safe area.

By adding the safe area margins as padding to the controls, we ensure
that the control's background flows into the non-safe area, while the
content item of the control (the toolbar or menu bar content) is kept
within the safe area.

Pick-to: 6.9
Change-Id: I5fad7394beaa01ae8ed142e4e2e42c5bffaab9fa
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>