qmlls: fix go to definition going to build folder
Amends b1ff6073a7247c921e20cd98b2e3297e7368681b that does not find
resource files of QML Modules that are not the main QML Module: use all
available .qrc files in the build directory when doing operations like
"go to definition".

Also search for hidden folders, because .qrc files tend to be nested
inside of .qt or .rcc folders.

Add a test when going to the defition of a QML file defined in a
different module.

Pick-to: 6.8 6.9
Fixes: QTCREATORBUG-31881
Fixes: QTBUG-131920
Change-Id: I751e2cd2d1a9867459226e4f884eb1da5c1f9436
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>