QtQml: Deal with platform-dependent results in ECMAScript test suite
For cases where the test expectations would need to cover multiple
conditions we finally throw the towel now and simply expect that the
test doesn't crash. Encoding the platforms with various degrees of
half-sane locale-dependent upper- and lower-casing into a domain
specific language is not something we're going to do here.

The nonICUFails expectation is accordingly rephrased as
"passesWithICU", which means we only expect it to survive otherwise, but
don't expect any particular result.

String/prototype/toLocaleUpperCase/special_casing.js now also fails on
Windows, so change its expectation to passesWithICU.

Change-Id: If53565de7e0df4a3d0ffea324a402b9f2f6b358f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>