Blacklist tst_FlickableInterop::pinchHandlerOnFlickable
We're seeing again an increase of failures in the test (including BFAIL
in some of the other test functions).

The corresponding bug has been closed as "Won't Do" which is probably
correct in the sense that this might need more work on
QQuickTouchUtils/Testlib instead of on the test side.

That doesn't change the fact that the test currenlty causes failed
integrations, bo blacklist it for now.

Task-number: QTBUG-95887
Change-Id: Ie9afcc34598c485679c02b3f29bf34b9f94dff51
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>