QtQml: Cache configuration parameters in type loader
The type loader should not access the engine from its thread. The
configuration values about debugging and cache modes are mostly
constant, except for the case where a debug client attaches later, when
the engine is already running.

Attaching a debugger or preview service without blocking the engine is
generally a bad idea because part of the application will already have
been loaded when the service is attached. That means some files will not
be replaced by the preview service and some files will run without debug
instructions, making it impossible to set break points.

To allow for the maximum flexibility, we read the configuration values
right before starting the thread. Since the thread is started on demand,
this is actually the last moment when we can safely do so.

Task-number: QTBUG-131721
Change-Id: Ida9039a5726a81c610bc96e5f473bd0875b74cd1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>