Break binding loop when binding implicit size to safe area margins
A common, and documented, pattern is to base the implicit
height of a control on its implicit content height, plus any
top or bottom padding.

With the introduction of safe areas another pattern emerges,
which is to bind the safe area margins to the top and bottom
padding of the control, to ensure that the content of the
control is positioned inside the safe area, if the control
itself is positioned in a way that overlaps a non-safe area.

Unfortunately this may lead to a binding loop, as increasing
the implicit height of the control due to the safe area padding
will potentially extend the control further into the non-safe
areas, depending on its position policy and anchoring, or if
the control overlaps both the top and bottom safe area margins,
in which case the control can never move the content item
into the safe area by moving the control itself up or down.

To short-circuit this binding loop, without forcing the user
to add complicated logic to their implicitHeight expressions,
we match the behavior when a control has an explicit height,
which is to shrink the control's content item. We do this by
detecting the situation, and reporting a smaller implicit
height in QQuickControl::implicitContentHeight, which then
results in a stable QQuickControl::implicitHeight.

Users who wish to override this behavior can bind to the
content item's implicitHeight directly, which will always
reflect the content item's real implicit height.

The exact same situation applies to the implicit content
width, so we use the same trick there as well.

Pick-to: 6.9
Change-Id: I3249fa877f0edd48cf411b060020111309268ebe
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>