ContextMenu: add more data rows to customContextMenu auto test
We need to ensure that it's possible to use ContextMenu on controls
like Pane, which accept mouse events to prevent them from going to
items beneath them.

customContextMenuOnControl.qml was using Label, which is derived from
QQuickText, not QQuickControl, so give Label its own file.

Also, add a test for the actual Control type just for completeness.

Task-number: QTBUG-132073
Pick-to: 6.9
Change-Id: I3521aa98a4aeec4848bf10c6c46f4e6e83bdb697
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>