QtQml: Allow basic name quoting in qmldir files
This adds support for quoted file names and for escaping quotes and
backslashes with '\'. It breaks support for file names that genuinely
start with a quote.

[ChangeLog][QtQml] You can now quote file names in qmldir files using
'"'. This is so that you can write file names with spaces in them. '\'
functions as an escape character, allowing you to escape quotes that are
part of file names and backslash itself.

Task-number: QTBUG-131916
Change-Id: Ia89e60153d3fafa273034f81da37166f30284c76
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>