Revert "Make ApplicationWindow's content item the fallback focused item on startup"
This reverts commit 71f687111b45a1d391c05fa26ef1cc72c75c9781.

The application window's content item having focus on startup
prior to 83d90845527799c78366c2b89e9c14bc35e88695 was a side-
effect of the item being a focus scope. In that situation any
parented item into the content item with focus will keep its
focus if no other item in the content item has focus at that
point.

However in the new setup the application window's content item
is not a focus scope (the root QQuickWindow content item is),
so giving it focus will prevent parented items from keeping
their focus. The workaround in the reverted patch of setting
the focus on completion instead of classBegin only worked for
the simple case where the focused item is part of the initial
object hierarchy. If an item is inserted after the app window
is completed this would still fail.

Change-Id: I129d8d8a7c048222d1af220bc209b4c7f20b97ca
Pick-to: 6.9
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>